requirements.txt
anthropic>=0.40.0
replicate>=0.34.0
youtube-transcript-api>=0.6.3
Pillow>=10.4.0
python-dotenv>=1.0.1
requests>=2.32.0
